Fix self links for repo collection by namespace
